pyproject.toml
[project]
name = "token-count"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llvmlite>=0.44.0",
    "numba==0.61.2",
    "numpy==2.2.6",
    "python-frontmatter>=1.1.0",
    "ruamel-yaml>=0.18.15",
    "tokenizers>=0.21.2",
    "transformers==4.53.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.16.1",
    "ruff>=0.12.1",
]

[tool.ruff.lint]
select = ["I"]

[tool.mypy]
exclude = "tokenizer/.*"

[[tool.mypy.overrides]]
module = ['tokenizer.tokenization_plamo']
disable_error_code = ["import-untyped", "override"]
